requirements.txt
Flask==3.0.2
spotipy==2.23.0
python-dotenv==1.2.1
groq>=0.4.0
flask-cors==6.0.1
psycopg2-binary==2.9.11
requests>=2.31.0